add decoding attn fp8

+++ src/decoding_attn_int8/block_info.h
@@ -16,13 +16,13 @@ struct QuantizatioInt8BlockInfo {
           sum_s_k(params.cu_seqlens_k[b]),
           actual_seqlen_k(params.cu_seqlens_k[b + 1] - sum_s_k) {}
 
-    inline __device__ size_t k_offset(const int row_stride, const int head_idx, const int head_stride,
-                                      const int dim_idx) const {
-        return (sum_s_k + s_k) * row_stride + head_idx * head_stride + dim_idx;
+    inline __device__ size_t k_offset(const size_t row_stride, const size_t head_idx, const size_t head_stride,
+                                      const size_t dim_idx) const {
+        return static_cast<size_t>(sum_s_k + s_k) * row_stride + head_idx * head_stride + dim_idx;
     }
 
-    inline __device__ size_t k_scale_offset(const int head_k, const int head_idx) const {
-        return (sum_s_k + s_k) * head_k + head_idx;
+    inline __device__ size_t k_scale_offset(const int head_k, const size_t head_idx) const {
+        return static_cast<size_t>(sum_s_k + s_k) * head_k + head_idx;
     }
 
     const int b;
@@ -44,17 +44,17 @@ struct DecodingInt8BlockInfo {
           row_shift(actual_seqlen_k - actual_seqlen_q),
           h_slope(1.0 / exp2f(8.0 * (h + 1) / params.h)) {}
 
-    inline __device__ size_t q_offset(const int row_stride, const int head_stride, const int dim_idx) const {
-        return sum_s_q * row_stride + h * head_stride + dim_idx;
+    inline __device__ size_t q_offset(const size_t row_stride, const size_t head_stride, const size_t dim_idx) const {
+        return static_cast<size_t>(sum_s_q) * row_stride + h * head_stride + dim_idx;
     }
 
-    inline __device__ size_t k_offset(const size_t seqlen_k, const int row_stride, const int head_stride,
-                                      const int dim_idx) const {
-        return (sum_s_k + seqlen_k) * row_stride + h_k * head_stride + dim_idx;
+    inline __device__ size_t k_offset(const size_t seqlen_k, const size_t row_stride, const size_t head_stride,
+                                      const size_t dim_idx) const {
+        return static_cast<size_t>(sum_s_k + seqlen_k) * row_stride + h_k * head_stride + dim_idx;
     }
 
     inline __device__ size_t k_scale_offset(const size_t seqlen_k, const int head_k) const {
-        return (sum_s_k + seqlen_k) * head_k + h_k;
+        return static_cast<size_t>(sum_s_k + seqlen_k) * head_k + h_k;
     }
 
     const int b;
